requirements.txt
jsonschema==3.2.0
matplotlib==3.3.4
networkx==2.5
numpy==1.17.4
PyYAML==5.3.1
